Define primary keys
