--- CourseMapping.Infrastructure/Configuration/SubjectConfiguration.cs
@@ -1,27 +1,0 @@
-using CourseMapping.Domain;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace CourseMapping.Infrastructure.Configuration;
-
-public class SubjectConfiguration : IEntityTypeConfiguration<Subject>
-{
-    public void Configure(EntityTypeBuilder<Subject> builder)
-    {
-        builder
-            .Property(s => s.Code)
-            .IsRequired();
-
-        builder
-            .Property(s => s.Name)
-            .IsRequired();
-        
-        builder
-            .Property(s => s.Description)
-            .IsRequired();
-        
-        builder
-            .Property(s => s.Level)
-            .IsRequired();
-    }
-}
