--- CourseMapping.Infrastructure/Configuration/UniversityConfiguration.cs
@@ -1,27 +1,0 @@
-using CourseMapping.Domain;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace CourseMapping.Infrastructure.Configuration;
-
-public class UniversityConfiguration : IEntityTypeConfiguration<University>
-{
-    public void Configure(EntityTypeBuilder<University> builder)
-    {
-        builder
-            .Property(u => u.Id)
-            .IsRequired();
-        
-        builder
-            .Property(u => u.Name)
-            .IsRequired();
-        
-        builder
-            .Property(u => u.Country)
-            .IsRequired();
-        
-        builder
-            .Property(u => u.Courses)
-            .IsRequired();
-    }
-}
